requirements.txt
asgiref==3.3.1
beautifulsoup4==4.4.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.44.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-google-translate==1.1
django-heroku==0.3.1
django-mathfilters==1.0.0
django-rest-framework==0.1.0
Django-Verify-Email==0.0.5
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
docutils==0.17
gunicorn==20.1.0
html-table-parser==0.1.0
idna==2.10
numpy==1.20.2
oauthlib==3.1.0
Pillow==8.1.0
playsound==1.2.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.2.4
urllib3==1.26.3
wfastcgi==3.0.0
whitenoise==5.2.0
